requirements.txt
# ── LLM e Agentes ─────────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── Persistência Redis ────────────────────────────────────────────────────────
redis>=5.0.0
langgraph-checkpoint-redis>=0.0.1

# ── Persistência relacional (staging de curadoria + observabilidade) ─────────
# SQLite como default para dev; Postgres (asyncpg) em produção. Ver ADR-023.
aiosqlite>=0.20.0
asyncpg>=0.29.0

# ── Qdrant (vetorial async) ──────────────────────────────────────────────────
qdrant-client>=1.11.0

# ── Busca web (câmbio) ────────────────────────────────────────────────────────
tavily-python>=0.3.0

# ── API (bridge FastAPI → LangGraph) ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── Interface legada (Streamlit, mantida para compatibilidade) ────────────────
streamlit>=1.40.0

# ── Utilitários ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testes e Simulador ────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0          # cliente HTTP async do simulador
rich>=13.0.0           # output colorido do simulador
